pyproject.toml
[project]
name = "micro-cti"
version = "0.1.0"
description = "OSINT app for collecting cybersecurity threat intelligence from microblogging platforms."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "atproto>=0.0.62",
    "beautifulsoup4>=4.14.2",
    "dictature>=0.14.0",
    "fastapi[standard]>=0.118.2",
    "feedparser>=6.0.12",
    "fuzzywuzzy[speedup]>=0.18.0",
    "lark>=1.3.0",
    "markdown>=3.9",
    "mastodon-py>=2.1.4",
    "openai>=2.2.0",
    "pyairtable>=3.2.0",
    "pydantic-ai>=1.0.17",
    "pydantic-ai-slim[mistral]>=1.0.17",
    "sqlalchemy>=2.0.48",
    "sqlmodel>=0.0.37",
    "telethon>=1.41.2",
    "tomli-w>=1.2.0",
]
